chore: remove monorepo for auth0 demo and improve instructions for testing locally
