--- demos/remote-mcp-auth0/.npmrc
@@ -1,4 +1,0 @@
-shamefully-hoist=true
-use-workspace-protocol=true
-link-workspace-packages = true
-prefer-workspace-packages = true
