+++ demos/remote-mcp-auth0/workers/mcp-auth0-oidc/.prettierrc
