+++ demos/remote-mcp-auth0/workers/todos-api/src/types/hono.ts
