+++ demos/remote-mcp-auth0/workers/todos-api/src/utils/http-exception.ts
