+++ demos/remote-mcp-auth0/workers/todos-api/wrangler.jsonc
@@ -11,6 +11,6 @@
     "enabled": true
   },
   "dev": {
-    "port": 8788
+    "port": 8789
   }
 }
